requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
python-dotenv==1.0.1

PyPDF2==3.0.1

faiss-cpu==1.10.0
numpy==1.26.4

openai==1.61.0

sqlglot==26.6.0

pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.1.0